requirements.txt
torch
torchvision
opencv-python
numpy
tqdm
Pillow
lxml
